serviceReward/requirements.txt
asgiref==3.5.2
Django==4.1.3
mysqlclient==2.1.1
sqlparse==0.4.3
tzdata==2022.7
djangorestframework
django-cors-headers
requests>=2.0.0,<3.0.0
pymongo